ibmvnic: Use flags when requesting FW statistics
Firmware accepts flags when the vnic client requests device statistics.
The first bit of the flag can be set to request physical port statistics
rather than logical port statistics. The second bit can be set to tell
FW to clear the statistics.

Previously, the byte for flags was memset to 0. However, in the future,
we will be using the clear statistics flag to reset the stats when
the device is being removed. Therefore, separate the request statistics
crq logic into a helper function and add an argument for specifying the
flags for the crq. Requesting statistics through ethtool behaves the
same as previous implementations in that it does not set any flags.

Signed-off-by: Nick Child <nnac123@linux.ibm.com>