ibmvnic: Reset FW stats during open
When loading a new ibmvnic device, reset the FW statistics for this virtual
device. This can be accomplished with the CRQ REQUEST_STATISTICS with the
CLEAR_STATS flag.

If this is not performed then it is possible that FW has existing
statistics for the virtual device. This scenario occurs when the ibmvnic
module is removed and reloaded. If the fw statistics are not cleared then
the fw statistics (requested through ethtool) will vary greatly versus the
netdevice statistics (requested through sysfs or ifconfig).

Signed-off-by: Nick Child <nnac123@linux.ibm.com>