ibmvnic: Include ethernet header in RX stats
There are two places which keep track of an ibmvnic devices statistics:
        1. In the netdevice structure
                - This is maintained within the driver and can be read
                through sysfs or with ifconfig
        2. In the firmware
                - This is maintained outside of the virtual environment but
                can be requested through ethtool.

Previously, the ibmvnic statistics in the netdevice structure were not
including the ethernet header in the rx total bytes statistic. On the
other hand, the FW does include these bytes. This could lead to a sizable
difference in the two rx total bytes numbers, especially when sending
smaller packets (as a 14 byte header would be a larger proportion of the
total size).

Therefore, include the ethernet header in the ibmvnic netdevice rx total
byte statistic. Since the call to eth_type_trans peels away the header,
just record the skb's length before calling this function.

Signed-off-by: Nick Child <nnac123@linux.ibm.com>