powerpc: Add version to install filenames
Rather than replacing the versionless vmlinux and System.map files,
copy to files with the version info appended.

Additionally, since executing the script is a last resort option,
inform the user about the missing `installkernel` command and the
location of the installation.

This work is adapted from `arch/s390/boot/install.sh`.

Signed-off-by: Nick Child <nnac123@linux.ibm.com>