ibmvnic: Only notify peers after qdiscs are assigned
If there is a call to netdev_notify_peers() immediately after a call
to netif_carrier_on() then there is a possibility that the ARPs from
netdev_notify_peers() will not be transmitted.

During any vnic transport event, an ibmvnic device will set the carrier
off. This schedules a work item which will assign noop_qdisc to the
transmit queues. Any packets attempted to be sent at this time will not
make it to the drivers xmit function.

Shortly after the vnic transport event, the ibmvnic device will set the
carrier up. Another, work item is queued which will reassign a default
qdisc to the transmit queues.

If the transport event was a failover or migration then the ibmvnic
device will call netdev_notify_peers() to inform its neighbors of any
address changes. This will transmit a gratuitous ARP across its network.
But, if the transmit queue is still assigned noop_qdisc then sending
the ARP will fail silently. As a result, active network connections
will hang until the peer sends an ARP requesting the new MAC.

Therefore, the ibmvnic device should wait until a proper qdisc is
assigned before calling netdev_notify_peers(). This will happen in
a workqueue and will reschedule itself until the qdisc is something
other than noop_qdisc or until the device is downed.

Signed-off-by: Nick Child <nnac123@linux.ibm.com>