ibmvnic: Track login requests in a list
It is possible for there to be multiple outstanding login requests.
Unfortunately, a LOGIN_RSP has no method for mapping the CRQ to a login
request. Therefore, there is no way to determine which login request has
been responded to. It is known that a login request sent after a
successful login request will return with a negative return code.

By storing all outstanding login requests in a list, the driver can
intuitively figure out which one has been responded to by checking if
the login response buffer has been filled in.

First, before sending a login response buffer to VIOS, poison the data.
Once a LOGIN_RSP is received, check which login response buffer has
accurate data (not poison). The most recent buffer sent with a valid
response buffer is the one related to the LOGIN_RSP event.

Additionally, get rid of the login_pending flag as the length of the
outstanding login requests list can provide the same purpose.

Lastly, let the release_login_buffer be in charge of dma unmapping the
buffers. Previously, DMA unmapping could be erroneously skipped if
a login request timed out. This would cause another call to send_login,
which would free the old buffer but not unmap.

Signed-off-by: Nick Child <nnac123@linux.ibm.com>