ibmvnic: Use hard_reset for CHANGE_PARAM reset
A CHANGE_PARAM reset is triggered from a userspace modification
to the devices capabilities. This can be from a request to change
the mtu, number of queues or ring size. Because this reset type
comes from userspace and is not a formal reset type defined in
PAPR, the process is essentially a full communication reset
with FW.

In fact, if you filter out everything that happens in do_reset()
during a CHANGE_PARAM reset, you are left with almost the exact
same function as do_hard_reset(). There is only one extra step
to tell the vnic server to close the link status.

Since do_reset() is a knotted mess of conditionals, call
do_hard_reset() instead during CHANGE_PARAM resets. Now CHANGE_PARAM
reset conditionals in do_reset() can be removed. This implements
the same functionality while improving overall code readability.

Signed-off-by: Nick Child <nnac123@linux.ibm.com>