ibmvnic: Exit early from unexpected state change during reset
There is a time when the rtnl_lock is dropped for a moment
in do_reset(). The intention was to allow for the linkwatch
events to have a change to fire. The effectiveness of this
approach is unclear but there is a definite possibility that
another process grabs the rtnl lock and changes the state of
the device.

Previously, these situations were handled by just pretending
that a parellel event was not happening. For obvious reasons
this is bad practice. If there is evidence that a parallel
task is modifying the driver, exit the reset and try again
later with fresh checks on the devices status. Additionally,
this conditional has never been true though testing and even
if it were to be hit then it is probably a sign of a larger
locking issue and should be patched properly.

Also, fix the loged message in these scenarios to be more
transparent with the user.

Signed-off-by: Nick Child <nnac123@linux.ibm.com>