ibmvnic: Remove unnecessary memsets
All buffers that the driver and FW share come with a length of
readable data value. Therefore clearing out this data after every
use is not really necessary. This data will get overwritten in the next
use and any older data that exceeds the length of the new data should
not be read because the length argument should bind consumer.

To save on cpu-cycles, remove the calls to memset memory to zero.

Signed-off-by: Your Name <you@example.com>
Signed-off-by: Nick Child <nnac123@linux.ibm.com>