hexdump: Prevent fallback to groupsize of 1 where possible
Previously, if len was not divisible by groupsize then groupsize of 1
was enforced. When printing a large buffer (calling hex_dump_to_buffer
several times) the non-uniform output can be annoying. For example,
below is output when printing a 31 byte buffer with 8 byte groupsize:
 6c6774732e737561 2e6d62692e736261   <- bytes 1-16 format correctly
 63 6f 6d 00 03 00 05 65 6e 76 33 00 00 00 00  <- bytes 17-31 different

To address this non-uniform output, make hex_dump_to_buffer use the
groupsize that caller provided (1, 2, 4 or 8). If there is a nonexact
len left to fill the group then just print a noncomplete group. In
the above example the new output is:
 6c6774732e737561 2e6d62692e736261
 636f6d0003000565 6e763300000000

Signed-off-by: Nick Child <nnac123@linux.ibm.com>