ibmveth: Modify replenish behaviour
Mostly code modification and some behavioural changes in the rx
replenish path. In prepration for a new hcall which can replenish
several buffers at once, move the descriptor generation and error
recovery to its own function ibmveth_fill_rx_descriptor.

The slight behavioural changes include pool->available and free_map
indexs being adjusted as soon as the descriptor is successfully generated.
Not only is this easier to manage, but it also makes more sense: as
soon as the index is removed from the free_map it should be considered
unavailable.

Previously, during hcall failure, cleanup was handled in the function.
This cleanup depended on the map tracking pointers being unmodified.
Since we are now modifying those pointers, the cleanup can simply be
given to ibmveth_remove_buffer_from_pool which does exactly what we
want; it unmaps and frees and sticks the index back into the list, it
even decrements pool->available. The only difficult thing is having
to track correlators since I don't see any simple way of getting the
pool and index information from the generated descriptor.

The goal of these changes are to make the introduction of batch sending
rx buffers easier.

Signed-off-by: Nick Child <nnac123@linux.ibm.com>